requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pyzmq==27.0.0
replicate==1.0.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
style==1.1.0
tabulate==0.9.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
update==0.0.1
urllib3==2.5.0
wcwidth==0.2.13
